requirements.txt
pytest
sphinx
grpcio
grpcio-tools
icecream
pytest
anthropic
jsonschema
rich
pydantic
pydantic-settings
requests 
pillow
ascii_magic
openai
notebook
ipywidgets
together
nbconvert
sphinx-material
